Remove `FeltBigInt`
